Make CInput wide character aware
This introduces an entirely new variant for get_char() as the semantics
are different: the wide version has two return values, the character
itself plus a flag to indicate if there it was a key code or a character
(get_char() used > 256 values for key codes, something that can;t be
done here).

Keep the old, non-Unicode aware get_char() for now, while we gradually
transition to the new version.